+++ HISTORY.rst
@@ -4,6 +4,7 @@
   in the GitLab CD/CI stages.
 - Update gevent from version 20.4.0 to 21.8.0.
 - Update gunicorn from version 20.0.4 to 20.1.0.
+- Add more explicit serializer read only fields.
 
 4.0.19 (2021-10-27)
 ===================